Updated readme and sample experience
Refactored some old code and moved it into a package ready to be cleaned up.
